Refactor type coercion logic to TypeCoercer
